Create script
This script will run nmap scan against a large file containing subnets with the ability to pick up where it left off if the connection down, etc. It will know exactly where it left off within any subnet it was currently scanning. It will create a directory and organize findings based on subnet.